Improved the process by getting ResourcePile.cs to keep HashSet of productivityUnits that effect it.
Improved ResourcePile.cs by getting it to store a HashSet of productivityUnits that are influencing its ProductionSpeed. Provided a RegisterProductivityUnit and a UnregisterProductivityUnit which Add and Remove the unit from this HashSet. They then RecalculateProductionSpeed which is limited by a call the Mathf.Min of the speed and 10.0f. For ProductivityUnit.cs I have removed the static variables and the ++ and --  keeping a count of instances, and made the ResourcePile responsible for keeping track of its own productivity by calling Unregister and Register ProductivityUnits.