Done way too much for one commit but it wasn't working so changed something else! See Below...
Renamed main.unity to challenge_main.unity
Renamed menu.unity to challenge_menu.unity
Created a BestScore.cs this contains the Instance of the BestScore that is accessible from both of the above. It also contains the HighScore which is defined in HighScore.cs as [System.Serializable]. An Awake function, SaveData also [System.Serializable], DisplayBestScore and methods for Saving and Loading the HighScore.
Renamed Menu.cs to ChallengeMenu.cs, added a [DefaultExecutionOrder(1000) and stuff for scene switching and exiting.
Amended ChallengeManager.cs replace to Text ScoreText with TMPUGUI ScoreText. Added a TMPUGUI BestScoreText. If the Escape key is pressed then return to the Menu. Added 'if (m_Points > BestScore.Instance.hightScore.Score) then replace the highScore with the BestScore.Instance.playerName and the m_Points before displaying the BestScoreText.text.
Now the game doesn't display the bricks at all :( Will look at again on Sunday, when I also hope to amend the Warehouse scene and get information about the ProductivityUnits displayed in the top right info box...